feat: allow inline static editor resize
