+++ frontend/packages/app/src/components/timesheet-row/components/inline-time-entry/timeEntryForm.tsx
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import { mergeClassNames as cn } from "@next-pms/design-system";
 import {
   Button,
   ErrorMessage,
@@ -79,10 +80,15 @@ export const TimeEntryForm = ({
                   onChange={(value) => field.handleChange(value)}
                   fixedMenu={false}
                   placeholder="Comment"
-                  editorClass="px-2 h-24 min-h-24 max-h-60 w-full min-w-64 max-w-[min(680px,calc(90vw-2rem))] resize prose-sm overflow-auto scrollbar-thin bg-white border rounded-md border-outline-gray-2 text-ink-gray-7 text-base leading-5.25"
+                  editorClass={cn(
+                    "px-2 h-24 min-h-24 max-h-60 resize prose-sm overflow-auto scrollbar-thin",
+                    "bg-surface-white border rounded-md border-outline-gray-2 text-ink-gray-7 text-base leading-5.25",
+                    "box-border w-full min-w-64 max-w-[min(680px,calc(90vw-2rem))]",
+                    "contain-[inline-size] wrap-anywhere **:max-w-full **:wrap-anywhere",
+                  )}
                 />
                 {field.state.value === "" ? (
-                  <span className="absolute text-sm align-middle right-5 bottom-1 text-ink-gray-4">
+                  <span className="absolute text-[12px] align-middle right-2 bottom-1 text-ink-gray-4">
                     ⌘+↵
                   </span>
                 ) : null}
